requirements.txt
streamlit
streamlit-chat
openai
python-dotenv
tqdm
jupyter
numpy
pandas
matplotlib
